Allow user to input file to be read and inputed
Since the querySchedule() function creates a new
file and writes a new schedule into it, we decided
to allow the user to input the file name. This way
after they run the program once, they can run it
again and see the free time and other information
about their newly inputed classes.